+++ ai_11/volodymyr_khailov/epic 7/Control test/result.txt
@@ -1,3 +1,1 @@
 Banner Size: 2.4 meters
-при написанні прізвища з лівого краю поля —>  до правого: 
-при написанні прізвища з правого краю поля <—  до лівого:
