--- ai_11/volodymyr_khailov/epic 7/Control test/task_first.cpp
@@ -1,77 +1,0 @@
-// Загальні відомості
-// В американській системі виміру довжин використовуються ярди, фути, дюйми. 1 Ярд = 3 фути. 1 Фут = 12 дюймів. 1 Дюйм = 2.54 см.
-// Футбольні поля мають можуть мати різну довжину
-// Довжина: мінімум 100 ярдів, максимум 130 ярдів.
-// Ширина: мінімум 50 ярдів, максимум 100 ярдів.
-
-// Довжина задається по варіанту 100 - 110 -120 - 130 для кожної групи інший розмір поля. ШІ -11  100 ярдів, ШІ -12 110 ярдів … і т.д.
-
-// Використовуючи Ваше прізвище розрахувати найменшу довжину, яка необхідна для розміщення банера з Вашим прізвищем.
-//  За умови, що прізвище можна записати лише в 1 ряд, а ширина букви(разом з інтервалом між символами) = кількості голосних букв у Вашому прізвищі * 0.8 метра.
-// На основі обчислених розмірів букви визначити символи, які будуть знаходитись на 3-х центральних(по довжині) метрах поля. 
-// Обчислення зробити для 2-х випадків:
-// 1) при написанні прізвища з лівого краю поля —>  до правого
-// 2) при написанні прізвища з правого краю поля <—  до лівого.
-// Результат записати.
-
-
-#include <iostream>
-
-using namespace std;
-
-int main() {
-    // Задане прізвище
-    string surname = "Khailov";  // Замініть "YourSurname" на своє прізвище
-
-    // Розмір букви і проміжку між символами (використана дійсна з подвійною точністю змінна)
-    double letterWidth = 0.8;  // метри
-    
-    // Кількість голосних букв у прізвищі (використана цілочисельна змінна)
-    int vowelsCount = 0;
-
-    // Підрахунок кількості голосних букв у прізвищі
-    for (char letter : surname) {
-        if (tolower(letter) == 'a' || tolower(letter) == 'e' || tolower(letter) == 'i' ||
-            tolower(letter) == 'o' || tolower(letter) == 'u') {
-            vowelsCount++;
-        }
-    }
-
-    // Розмір банера (використана дійсна змінна)
-    double bannerSize = vowelsCount * letterWidth;
-
-    // Виведення результату
-    cout << "Banner Size: " << bannerSize << " meters" << endl;
-
-    // Визначення символів на центральних трьох метрах (використано умовний оператор та математичні операції)
-
-    // 1 Ярд = 3 фути. 1 Фут = 12 дюймів. 1 Дюйм = 2.54 см.
-    double yardsToMeters = 0.9144;  // Коефіцієнт переведення з ярдів в метри
-    double fieldLengthInYards = 100.0;  // Довжина поля в ярдах
-    double fieldLengthInMeters = fieldLengthInYards * yardsToMeters;  // Переведення в метри
-
-    int centralMeters = 3;  // Кількість центральних метрів
-
-    // Визначення позицій символів для лівого написання (використано цілочисельні змінні та одновимірний масив)
-    int startLeft = (fieldLengthInYards * 3) / 2 - centralMeters;
-    int endLeft = startLeft + centralMeters;
-
-    // Визначення позицій символів для правого написання (використано цілочисельні змінні та одновимірний масив)
-    int startRight = (fieldLengthInYards * 3) / 2;
-    int endRight = startRight + centralMeters;
-
-    // Виведення результату
-    cout << "при написанні прізвища з лівого краю поля —>  до правого: ";
-    for (int i = startLeft; i < endLeft; i++) {
-        cout << surname[i % surname.length()] << " ";
-    }
-    cout << endl;
-
-    cout << "при написанні прізвища з правого краю поля <—  до лівого: ";
-    for (int i = endRight - 1; i >= startRight; i--) {
-        cout << surname[i % surname.length()] << " ";
-    }
-    cout << endl;
-
-    return 0;
-}
